separated nuget json parser into own module;
